src/is-robot/pyproject.toml
[tool.poetry]
name = "is-robot"
version = "0.1.0"
description = ""
authors = ["Luiz Carlos Cosmi Filho <luizcarloscosmifilho@gmail.com>"]
readme = "README.md"
packages = [{include = "is_robot"}]

[tool.poetry.dependencies]
python = "^3.8"
is-wire = {git = "https://github.com/labviros/is-wire-py.git", rev = "add_communication_tracing"}
vine = "1.3.0"
is-msgs = "1.1.10"
numpy = "1.22.4"
python-dateutil = "2.8.0"
opencensus-ext-zipkin = "0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[tool.poetry.scripts]
is-robot = 'is_robot.service:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
